organize convolist better
