+++ src/components/Table/stories/ConvoList/ConvoList.customCells.js
@@ -1,0 +1,29 @@
+import React from 'react'
+import { TAG_COLORS } from './ConvoList.constants'
+import {
+  TagUI,
+  SubjectUI,
+  PreviewUI,
+  ConversationCellUI,
+} from './ConvoList.css'
+
+export function ConversationCell({ subject, preview, tags, row }) {
+  return (
+    <ConversationCellUI>
+      <div>
+        {tags
+          ? tags.map(tag => (
+              <TagUI
+                color={TAG_COLORS[tag.color]}
+                key={`${tag.name}_${row.id}`}
+              >
+                {tag.name}
+              </TagUI>
+            ))
+          : null}
+        <SubjectUI>{subject}</SubjectUI>
+      </div>
+      <PreviewUI>{preview}</PreviewUI>
+    </ConversationCellUI>
+  )
+}
