+++ src/components/Table/stories/ConvoList/convoList.utils.js
@@ -1,0 +1,6 @@
+export function formatDate(date) {
+  return new Date(date).toLocaleDateString('en-US', {
+    month: 'short',
+    day: 'numeric',
+  })
+}
